TypeScript types for Recoil
