+++ types/recoil/lib/core/RecoilRoot.d.ts
@@ -1,0 +1,11 @@
+import * as React from 'react';
+import { RecoilState } from './recoilValue';
+
+export type RecoilRootProps = {
+    initializeState?: (options: {
+        set: <T>(recoilVal: RecoilState<T>, newVal: T) => void;
+        setUnvalidatedAtomValues: (atomMap: Map<string, unknown>) => void;
+    }) => void;
+};
+
+export const RecoilRoot: React.FC<RecoilRootProps>;
