+++ types/recoil/lib/core/node.d.ts
@@ -1,0 +1,1 @@
+export class DefaultValue {}
