+++ types/recoil/lib/core/selector.d.ts
@@ -1,0 +1,32 @@
+import { RecoilValue, RecoilState, RecoilValueReadOnly } from './recoilValue';
+import { DefaultValue } from './node';
+
+export type GetRecoilValue = <T>(recoilVal: RecoilValue<T>) => T;
+export type SetRecoilState = <T>(
+    recoilVal: RecoilState<T>,
+    newVal: T | DefaultValue | ((prevValue: T) => T | DefaultValue),
+) => void;
+
+export type ResetRecoilState = <T>(recoilVal: RecoilState<T>) => void;
+
+export type ReadOnlySelectorOptions<T> = {
+    key: string;
+    get: (opts: { get: GetRecoilValue }) => Promise<T> | RecoilValue<T> | T;
+
+    // cacheImplementation_UNSTABLE?: CacheImplementation<Loadable<T>>,
+    dangerouslyAllowMutability?: boolean;
+};
+
+export type ReadWriteSelectorOptions<T> = ReadOnlySelectorOptions<T> & {
+    set: (
+        opts: {
+            set: SetRecoilState;
+            get: GetRecoilValue;
+            reset: ResetRecoilState;
+        },
+        newValue: T | DefaultValue,
+    ) => void;
+};
+
+export function selector<T>(options: ReadWriteSelectorOptions<T>): RecoilState<T>;
+export function selector<T>(options: ReadOnlySelectorOptions<T>): RecoilValueReadOnly<T>;
