+++ types/recoil/tslint.json
@@ -1,1 +1,7 @@
-{ "extends": "dtslint/dt.json" }
+{
+    "extends": "dtslint/dt.json",
+    "rules": {
+        "interface-over-type-literal": false,
+        "no-unnecessary-generics": false
+    }
+}
